移除多余的代码

+++ handler/folder.go
@@ -175,7 +175,6 @@ func (f *folderHandler) DeleteSource(c *gin.Context) {
 	deleteMultiObjects := model2.DeleteMultiObjects{
 		Quiet: false, //详细和静默模式，设置为 true 的时候，只返回删除错误的文件列表，设置为 false 的时候，成功和失败的文件列表都返回
 	}
-	var fileHashList []string
 	authId := middleware.UserId(c)
 	// 删除指定的文件
 	if len(l.FileIds) > 0 {
@@ -205,20 +204,17 @@ func (f *folderHandler) DeleteSource(c *gin.Context) {
 			deleteMultiObjects.Append(model2.DeleteObject{Key: hash[:2] + "/" + hash[2:]})
 		}
 	}
-	for _, hash := range fileHashList {
-		deleteMultiObjects.Append(model2.DeleteObject{
-			Key: hash[:2] + "/" + hash[2:]},
-		)
-	}
-	deleteRequest := &model2.DeleteMultiObjectsRequest{
-		Bucket:        f.bucketName,
-		DelectObjects: &deleteMultiObjects,
-	}
-	fmt.Println(f.bucketName)
-	_, err := f.nosClient.DeleteMultiObjects(deleteRequest)
-	if err != nil {
-		_ = c.Error(errors.BadRequest(fmt.Sprintf("删除文件失败: %+v", err), err))
-		return
+
+	if len(deleteMultiObjects.Objects) > 0 {
+		deleteRequest := &model2.DeleteMultiObjectsRequest{
+			Bucket:        f.bucketName,
+			DelectObjects: &deleteMultiObjects,
+		}
+		_, err := f.nosClient.DeleteMultiObjects(deleteRequest)
+		if err != nil {
+			_ = c.Error(errors.BadRequest(fmt.Sprintf("删除文件失败: %+v", err), err))
+			return
+		}
 	}
 	c.Status(http.StatusNoContent)
 }
